Extract VmpResolver and fix Rubocop offences
- Extract resolve_vmp/de_brand/amp_or_ampp? out of NhsDmd::Search into
  a new NhsDmd::VmpResolver service to satisfy Metrics/ClassLength
- Inline the vmp_result let in the search spec context to drop the
  memoized helper count below the RSpec/MultipleMemoizedHelpers limit
